Dockerized

+++ src/main/java/com/nm/novamart/Repository/OrderItemRepository.java
@@ -3,5 +3,5 @@ package com.nm.novamart.Repository;
 import com.nm.novamart.Entity.OrderItems;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<OrderItems, Integer> {
+public interface OrderItemRepository extends JpaRepository<OrderItems, Long> {
 }
